Cleanup deprecated files and refactor imports (#280)
* Remove deprecated Pipeline class from the codebase.

This class was previously marked for deprecation and has been fully replaced by the Flow class. Its removal streamlines the code and encourages users to adopt the new implementation directly.

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Remove prompts.py file containing various chat templates

We have migrated to chat completions which makes this file irrelevant

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Remove registry.py file, which contained the BlockRegistry and PromptRegistry classes.

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Remove unrelevant file, pushed by mistake

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Remove unused scripts

deleted __init__.py and ruff.sh files to streamline the codebase.

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Refactor import statement in __init__.py to use relative import for BlockRegistry

Signed-off-by: shiv <shivchander.s30@gmail.com>

* Remove deprecated test file for the Pipeline class.

This file contained tests that are no longer relevant following the removal of the Pipeline class from the codebase, streamlining the test suite and encouraging the use of the new Flow implementation.

Signed-off-by: shiv <shivchander.s30@gmail.com>

---------

Signed-off-by: shiv <shivchander.s30@gmail.com>